refactor: Change folder structure
