+++ src/pages/analysis/Analysis.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { withStyles } from '@material-ui/core/styles';
 import Paper from '@material-ui/core/Paper';
 import Grid from '@material-ui/core/Grid';
-import Profile from '../Profile';
+import Profile from '../../components/Profile';
 
 const styles = (theme) => ({
   root: {
