+++ src/pages/difficulty/Difficulty.js
